Corrección de menú superior para tema claro y oscuro
