+++ app/src/main/java/com/example/whereyou/GruposActivity.kt
@@ -3,6 +3,8 @@ package com.example.whereyou
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.view.Menu
+import android.view.MenuItem
 import android.widget.ArrayAdapter
 import android.widget.Toast
 import com.example.whereyou.databinding.ActivityGruposBinding
@@ -13,6 +15,20 @@ import com.google.firebase.auth.FirebaseAuth
 class GruposActivity : AppCompatActivity() {
     private lateinit var auth: FirebaseAuth
     private lateinit var binding: ActivityGruposBinding
+    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+        menuInflater.inflate(R.menu.main_menu, menu)
+        return true
+    }
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        val clicked = item.itemId
+        if(clicked == R.id.menuLogOut){
+
+            val i = Intent(this, MainActivity::class.java)
+            intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
+            startActivity(i)
+        }
+        return super.onOptionsItemSelected(item)
+    }
 
     private val array = arrayOf("Familia", "Amigos", "Viaje a medellin", "Parche del viernes")
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -24,13 +40,6 @@ class GruposActivity : AppCompatActivity() {
         binding.GAListaGrupos.adapter = adapter
 
         binding.GAListaGrupos.setOnItemClickListener { adapterView, view, i, l -> startActivity(Intent(baseContext,MapaGrupoActivity::class.java))}
-
-        binding.GAMenuLogOut.setOnClickListener {
-            auth.signOut()
-            val i = Intent(this, LoginActivity::class.java)
-            intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
-            startActivity(i)
-        }
 
         binding.GANuevoContacto.setOnClickListener {
             startActivity(Intent(baseContext,CrearGrupoActivity::class.java))
