+++ app/src/main/java/com/example/whereyou/NotificacionesActivity.kt
@@ -3,6 +3,8 @@ package com.example.whereyou
 import DisplayChat
 import android.content.Intent
 import android.os.Bundle
+import android.view.Menu
+import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
 import com.example.whereyou.databinding.ActivityNotificacionesBinding
 import com.example.whereyou.datos.Contacto
@@ -10,6 +12,20 @@ import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class NotificacionesActivity : AppCompatActivity() {
     private lateinit var binding : ActivityNotificacionesBinding
+    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+        menuInflater.inflate(R.menu.main_menu, menu)
+        return true
+    }
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        val clicked = item.itemId
+        if(clicked == R.id.menuLogOut){
+
+            val i = Intent(this, MainActivity::class.java)
+            intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
+            startActivity(i)
+        }
+        return super.onOptionsItemSelected(item)
+    }
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityNotificacionesBinding.inflate(layoutInflater)
@@ -44,10 +60,6 @@ class NotificacionesActivity : AppCompatActivity() {
         binding.GAPerfil.setOnClickListener {
             startActivity(Intent(baseContext,PerfilActivity::class.java))
         }
-        binding.GAMenuLogOut.setOnClickListener {
-            val i = Intent(this, LoginActivity::class.java)
-            intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
-            startActivity(i)
-        }
+
     }
 }
