+++ app/src/main/res/layout/activity_crear_grupo.xml
@@ -6,52 +6,16 @@
     android:layout_height="match_parent"
     tools:context=".CrearGrupoActivity">
 
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbarNotificaciones"
-        android:layout_width="match_parent"
-        android:layout_height="60dp"
-        android:background="@color/azul3D"
-        android:minHeight="?attr/actionBarSize"
-        android:theme="?attr/actionBarTheme"
-        app:layout_constraintBottom_toTopOf="@+id/linearL"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-
-    <ImageButton
-        android:id="@+id/GAMenuLogOut"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:background="@android:color/transparent"
-        android:padding="8dp"
-        android:src="@drawable/baseline_logout_24"
-        app:layout_constraintBottom_toBottomOf="@id/toolbarNotificaciones"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="@id/toolbarNotificaciones" />
-
-    <TextView
-        android:id="@+id/titleText"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="16dp"
-        android:layout_marginStart="16dp"
-        android:gravity="left"
-        android:text="@string/crear_grupo"
-        android:textColor="#000000"
-        android:textSize="18sp"
-        android:textStyle="bold"
-        tools:ignore="MissingConstraints" />
-
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/constraintL"
         android:layout_width="0dp"
         android:layout_height="70dp"
         android:background="@color/gray"
-        app:layout_constraintBottom_toTopOf="@+id/linearL"
+        app:layout_constraintBottom_toTopOf="@+id/listaContactos"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.0">
 
         <androidx.appcompat.widget.SearchView
@@ -71,33 +35,6 @@
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 
-    <LinearLayout
-        android:id="@+id/linearL"
-        android:layout_width="0dp"
-        android:layout_height="60dp"
-        android:layout_marginTop="8dp"
-        android:orientation="horizontal"
-
-
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/toolbarNotificaciones">
-
-
-
-    </LinearLayout>
-
-    <View
-        android:id="@+id/separador"
-        android:layout_width="0dp"
-        android:layout_height="2dp"
-        android:layout_marginStart="20dp"
-        android:layout_marginEnd="70dp"
-        android:background="@color/black"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/linearL" />
-
     <ListView
         android:id="@+id/listaContactos"
         android:layout_width="0dp"
@@ -105,7 +42,7 @@
         app:layout_constraintBottom_toTopOf="@+id/GA_barraPrincipal"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/separador" />
+        app:layout_constraintTop_toBottomOf="@+id/constraintL" />
 
     <ImageButton
         android:id="@+id/imageButton"
