+++ app/src/main/res/layout/activity_grupos.xml
@@ -7,47 +7,6 @@
     tools:context=".GruposActivity">
 
 
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbarNotificaciones"
-        android:layout_width="match_parent"
-        android:layout_height="60dp"
-        android:background="@color/azul3D"
-        android:minHeight="?attr/actionBarSize"
-        android:theme="?attr/actionBarTheme"
-        app:layout_constraintBottom_toTopOf="@+id/GA_buscador"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ImageButton
-        android:id="@+id/GAMenuLogOut"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:background="@android:color/transparent"
-        android:padding="8dp"
-        android:src="@drawable/baseline_logout_24"
-        app:layout_constraintBottom_toBottomOf="@id/toolbarNotificaciones"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/Perfil"
-        app:layout_constraintTop_toTopOf="@id/toolbarNotificaciones" />
-
-    <TextView
-        android:id="@+id/Perfil"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="16dp"
-        android:layout_marginStart="16dp"
-        android:gravity="left"
-        android:text="@string/grupos"
-        android:textColor="#000000"
-        android:textSize="18sp"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toBottomOf="@id/toolbarNotificaciones"
-        app:layout_constraintEnd_toStartOf="@id/GAMenuLogOut"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@id/toolbarNotificaciones" />
 
 
 
@@ -57,11 +16,11 @@
         android:layout_height="50dp"
         android:layout_marginBottom="6dp"
         android:inputType="text"
-        android:text="Nombre de grupo o persona"
+        android:hint="Nombre de grupo o persona"
         app:layout_constraintBottom_toTopOf="@+id/GA_listaGrupos"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/toolbarNotificaciones" />
+        app:layout_constraintTop_toTopOf="parent" />
 
     <ListView
         android:id="@+id/GA_listaGrupos"
