+++ app/src/main/res/layout/activity_mapa_grupo.xml
@@ -280,16 +280,15 @@
         app:layout_constraintBottom_toTopOf="@+id/peopleMenu"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/toolbarNotificaciones">
-
+        app:layout_constraintTop_toTopOf="parent">
 
 
         <ImageView
             android:id="@+id/mapaViewGrupo"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:src="@drawable/mapagps"
             android:scaleType="fitXY"
+            android:src="@drawable/mapagps"
             app:layout_constraintBottom_toTopOf="@+id/peopleMenu"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
@@ -298,9 +297,9 @@
         <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:gravity="bottom"
             android:layout_gravity="bottom"
             android:background="@color/white"
+            android:gravity="bottom"
             android:orientation="vertical">
 
             <TextView
