+++ app/src/main/res/layout/activity_perfil.xml
@@ -6,48 +6,6 @@
     android:layout_height="match_parent"
     tools:context=".PerfilActivity">
 
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbarNotificaciones"
-        android:layout_width="match_parent"
-        android:layout_height="60dp"
-        android:background="@color/azul3D"
-        android:minHeight="?attr/actionBarSize"
-        android:theme="?attr/actionBarTheme"
-        app:layout_constraintBottom_toTopOf="@+id/botonCambiarFoto"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-
-    <ImageButton
-        android:id="@+id/GAMenuLogOut"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:background="@android:color/transparent"
-        android:padding="8dp"
-        android:src="@drawable/baseline_logout_24"
-        app:layout_constraintBottom_toBottomOf="@id/toolbarNotificaciones"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/Perfil"
-        app:layout_constraintTop_toTopOf="@id/toolbarNotificaciones" />
-
-    <TextView
-        android:id="@+id/Perfil"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="16dp"
-        android:layout_marginStart="16dp"
-        android:gravity="left"
-        android:text="@string/perfil"
-        android:textColor="#000000"
-        android:textSize="18sp"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toBottomOf="@id/toolbarNotificaciones"
-        app:layout_constraintEnd_toStartOf="@id/GAMenuLogOut"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@id/toolbarNotificaciones" />
-
 
     <LinearLayout
         android:id="@+id/CA_barraPrincipal"
@@ -170,16 +128,14 @@
     <ImageButton
         android:id="@+id/botonCambiarFoto"
         android:layout_width="400dp"
-        android:layout_height="150dp"
-        android:layout_marginTop="32dp"
+        android:layout_height="300dp"
+        android:background="@color/white"
+        android:scaleType="fitCenter"
+        android:src="@drawable/person_login"
         app:layout_constraintBottom_toTopOf="@+id/textView3"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/toolbarNotificaciones"
-        android:src="@drawable/person_login"
-        android:scaleType="fitCenter"
-        android:background="@color/white"
-        />
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
         android:id="@+id/textView3"
@@ -215,15 +171,6 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/textView4">
 
-        <TextView
-            android:id="@+id/textView5"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="16dp"
-            android:layout_marginTop="16dp"
-            android:text="Información"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
 
         <TextView
             android:id="@+id/textView6"
