+++ app/src/main/res/values/themes.xml
@@ -4,6 +4,10 @@
         <!-- Customize your light theme here. -->
         <item name="colorPrimary">@color/azul68</item>
     </style>
+    <style name="AppTheme.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
+    <style name="MyMenuItemText" parent="Widget.AppCompat.PopupMenu">
+        <item name="android:textColor">@color/black</item>
+    </style>
 
     <style name="Theme.WhereYou" parent="Base.Theme.WhereYou" />
 </resources>
